requirements_host.txt
mss
netifaces
numpy
opencv-python
Pillow
pyperclip
pypiwin32
PyQt5
PyQt5-sip
PyQt5-stubs
pytesseract
pyttsx3
pywin32
pyWinhook
regex
requests
xlrd==1.2.0
matplotlib
cryptography
discord.py==1.7.3
PyNaCl
gTTS
pymongo
dnspython
